ptits modifs

+++ libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: jubaldo <jubaldo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 17:33:24 by jubaldo           #+#    #+#             */
-/*   Updated: 2024/05/18 17:33:51 by jubaldo          ###   ########.fr       */
+/*   Updated: 2024/06/02 18:12:09 by jubaldo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,8 +48,8 @@ static char	*word_dup(const char *str, int start, int finish)
 
 char	**ft_split(char const *s, char c)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 	int		index;
 	char	**split;
